+++ src/WebApiBolierplate/Business.Lib/HelperLib.cs
@@ -8,53 +8,29 @@ namespace Business.Lib
     {
         #region [Security]
 
-        public string HashBCrypt(string plainText)
-        {
-            return helper.Hash.HashBCrypt(plainText);
-        }
+        public string HashBCrypt(string plainText) => helper.Hash.HashBCrypt(plainText);
 
-        public bool VerifyBCrypt(string plainText, string hash)
-        {
-            return helper.Hash.VerifyBCrypt(plainText, hash);
-        }
+        public bool VerifyBCrypt(string plainText, string hash) => helper.Hash.VerifyBCrypt(plainText, hash);
 
-        public string EncryptString(string clearText)
-        {
-            return helper.Encryption.EncryptString(clearText);
-        }
+        public string EncryptString(string clearText) => helper.Encryption.EncryptString(clearText);
 
-        public string DecryptString(string cipherText)
-        {
-            return helper.Encryption.DecryptString(cipherText);
-        }
+        public string DecryptString(string cipherText) => helper.Encryption.DecryptString(cipherText);
 
         #endregion [Security]
 
         #region [Utilities]
 
-        public int GenRandomNumber(int min, int max)
-        {
-            return helper.Rando.GenRandomNumber(min, max);
-        }
+        public int GenRandomNumber(int min, int max) => helper.Rando.GenRandomNumber(min, max);
 
-        public string GenRandomChar(int length)
-        {
-            return helper.Rando.GenRandomChar(length, CharSet.AlphaNumeric);
-        }
+        public string GenRandomChar(int length) => helper.Rando.GenRandomChar(length, CharSet.AlphaNumeric);
 
         #endregion [Utilities]
 
         #region [Compression]
 
-        public string CompressToGzipString(string plaintText)
-        {
-            return helper.GZip.CompressToString(plaintText, Encoding.UTF8);
-        }
+        public string CompressToGzipString(string plaintText) => helper.GZip.CompressToString(plaintText, Encoding.UTF8);
 
-        public string DecompressGzipString(string compressedString)
-        {
-            return helper.GZip.DecompressString(compressedString, Encoding.UTF8);
-        }
+        public string DecompressGzipString(string compressedString) => helper.GZip.DecompressString(compressedString, Encoding.UTF8);
 
         #endregion [Compression]
     }