+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.0] - 2019-05-09
+### Added
+ - Documentation for encryption helper
+
+### Changed
+ - Renamed CSV Logger
+ - Optimised code in helpersLib
+
 ## [3.1.0] - 2019-05-08
 ### Changed
  - Added sub classes to helper